pyproject.toml
[project]
name = "bivium"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "openai-agents>=0.13.4",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "shapely>=2.1.2",
    "uvicorn>=0.43.0",
    "uvloop>=0.22.1",
]

[dependency-groups]
dev = [
    "arize-phoenix-client>=2.3.0",
    "arize-phoenix-otel>=0.15.0",
    "openinference-instrumentation-openai-agents>=1.4.1",
    "pre-commit>=4.5.1",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.9",
]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["src/backend/tests"]

[tool.pyright]
typeCheckingMode = "strict"
include = ["src/backend"]
venvPath = "."
venv = ".venv"
